requirements.txt
ebaysdk
requests
python-dotenv
cachetools